+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Copy project, desc or ext. id to the clipboard on click over
+
 ## [v0.6.1] - 2024-05-12
 
 ### Changed